Switch floors and show correct floors in preview
It's now possible to switch floor by clicking on on e of the preview boxes in the building view. This takes you to the correct floor, updates and re-colours the different SVGs shown in the building view. The correct floors are now shown when the user enters the building view by reading from the JSON house structures file that contains information about this and how the different floors are connected